update: move action and reducer types to types files
