+++ src/redux/ducks/CardList/types.tsx
@@ -1,0 +1,4 @@
+export interface ListDataProps {
+    id: string;
+    listTitle: string;
+  };
